Fix floating-point comparison
Fix stuck movement, keeping `EPSILON_POS` and using proper
floating-point comparison.

+++ src/cdogs/utils.c
@@ -392,7 +392,7 @@ struct vec2 CalcClosestPointOnLineSegmentToPoint(
 	// P(t) = l1 + t(l2 - l1)
 	// Projection of point p on line is
 	// t = ((p.x - l1.x)(l2.x - l1.x) + (p.y - l1.y)(l2.y - l1.y)) / ||l2 - l1||^2
-	const float lineDistanceSquared = (int)svec2_distance_squared(l1, l2);
+	const float lineDistanceSquared = svec2_distance_squared(l1, l2);
 	// Early exit since same point means 0 distance, and div by 0
 	if (lineDistanceSquared == 0)
 	{
